Make U2U admin responsive: step 1 (affiliate partners + general)
